user_managemenet/TokenMiddleware.py
from django.contrib.auth.models import AnonymousUser
from rest_framework.authtoken.models import Token
from channels.db import database_sync_to_async
from channels.middleware import BaseMiddleware
import jwt
from django.conf import settings

@database_sync_to_async
def get_user(token_key):
    try:
        token = Token.objects.get(key=token_key)
        return token.user

    except Token.DoesNotExist:
        return AnonymousUser()

class TokenAuthMiddleware(BaseMiddleware):
    def __init__(self, inner):
        super().__init__(inner)

    async def __call__(self, scope, receive, send):
        # import pdb;pdb.set_trace()
        try:
            token_key = (dict((x.split('=') for x in scope['query_string'].decode().split("&")))).get('token', None)
            # import pdb;pdb.set_trace()
            payload = jwt.decode(token_key,settings.SECRET_KEY, algorithms='HS512')
            token_key=payload['token']
        except ValueError:
            token_key = None
        scope['user'] = AnonymousUser() if token_key is None else await get_user(token_key)
        # import pdb;pdb.set_trace()
        return await super().__call__(scope, receive, send)