requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-cors-headers==3.11.0
django-filter==21.1
django-rest-passwordreset==1.2.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-datatables==0.7.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyparsing==3.0.8
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
redis==4.3.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.0.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
